+++ apps/web/src/app/[locale]/(auth)/auth/logout/route.ts
