+++ apps/web/src/app/[locale]/(auth)/login/login-wizard/login-otp.tsx
@@ -14,11 +14,11 @@ import { ChevronLeftIcon } from "lucide-react";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
 
-import { InputOTP } from "@/app/[locale]/(unauthenticated)/login/components/input-otp";
+import { InputOTP } from "@/app/[locale]/(auth)/login/components/input-otp";
 import {
   useLoginWizard,
   useLoginWizardProps,
-} from "@/app/[locale]/(unauthenticated)/login/login-wizard/login-wizard";
+} from "@/app/[locale]/(auth)/login/login-wizard/login-wizard";
 import { useTranslation } from "@/app/i18n/client";
 import { Trans } from "@/components/trans";
